requirements.txt
# Bot Telegram
python-telegram-bot[job-queue]==20.7
nest-asyncio==1.5.8

# Spotify API
spotipy==2.23.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Environment Variables
python-dotenv==1.0.0

# Optional: Database upgrade (instead of CSV)
# sqlite3 is built-in to Python
# For PostgreSQL:
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.23